Rittenhouse/requirements.txt
Flask
Jinja2
mysql-connector-python
petl